Replaced `MapPage` with `Homepage`
